Working on the APIs and the frontend

+++ bacheca/src/main/java/com/homates/bacheca/repo/DocumentRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 
 
-public interface DocumentRepository extends CrudRepository<Document, Long>{
+public interface DocumentRepository extends CrudRepository<Document, Integer>{
     List<Document> findByCategory(String category);
 }
